requirements.txt
#pip install -r requirements.txt

farm-haystack[inference, opensearch, file-conversion]
fastapi
uvicorn[standard]